Move z_schema to controller/helpers/validator
