--- framework/src/modules/http_api/helpers/z_schema.js
@@ -1,1 +1,0 @@
-../../chain/helpers/z_schema.js
